CourseTableのCoreを分離
