+++ web/src/components/course/components/CoursesTableCore/styles.module.css
